requirements.txt
streamlit==1.32.0
openai==1.13.3
pillow==10.2.0
python-dotenv==1.0.0
tiktoken==0.5.2
httpx==0.27.2 